requirements.txt
smc-python>=0.3.8
ipaddress==1.0.17
boto3==1.4.1
PyYAML==3.12